Factor out toggle smooth
